app/requirements.txt
fastapi
uvicorn[standard]
jinja2
langchain
transformers
sentence-transformers
torch
accelerate
faiss-cpu
python-multipart
sqlite-utils
pypdf
python-docx